requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
einops==0.8.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
ftfy==6.3.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open-clip-torch==2.23.0
packaging==24.2
pandas==2.2.3
path==17.1.0
patsy==1.0.1
pillow==11.1.0
protobuf==5.29.3
pycocotools==2.0.8
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-image==0.25.2
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
statsmodels==0.14.4
sympy==1.13.1
tifffile==2025.3.30
timm==1.0.14
tokenizers==0.21.0
torch==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
transformers==4.48.3
triton==3.1.0
tueplots==0.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
